Updated
The old version would sometimes not split at all. This was now fixed. Also a new variable was introduced to make the split no longer occur when the fadeout started, but will now split once the loading screen appears.
Next step is to allow proper timing and settings for IL's.